@@ -4,6 +4,9 @@
     {
         [Required]
         [StringLength(8, MinimumLength = 4)]
-        public string NewPassword { get; set; } = null!;
+        public string? NewPassword { get; set; }
+
+        [Required]
+        public string? OldPassword { get; set; }
     }
 }
