@@ -2,7 +2,7 @@
 {
     public static class HttpContextExtensions
     {
-        public static int GetCurrentUserId(this HttpContext context)
-            => int.Parse(context.User.FindFirst(ClaimTypes.NameIdentifier)!.Value);
+        public static int? GetCurrentUserId(this HttpContext context)
+            => int.TryParse(context.User?.FindFirst(ClaimTypes.NameIdentifier)?.Value, out int userId) ? userId : null;
     }
 }
