@@ -4,7 +4,8 @@
     {
         public List<Deposit> GetUserDeposits(int userId);
         public List<Deposit> GetAllDeposits();
-        public bool ApproveTransaction(int txId);
-        public bool DeclineTransaction(int txId);
+        public (ServiceResult Result, string Message) ApproveTransaction(int txId);
+        public (ServiceResult Result, string Message) DeclineTransaction(int txId);
+        public Deposit? GetTransactionById(int txId);
     }
 }
