@@ -2,7 +2,7 @@
 {
     public interface IUserBalanceService
     {
-        public List<UserBalance>? GetUserBalances(int userId);
+        public (bool Result, List<UserBalance>? Balances) GetUserBalances(int userId);
         public void AddUserBalance(int userId, string currency, decimal amount);
     }
 }
