@@ -17,27 +17,35 @@
         public List<Deposit> GetUserDeposits(int userId)
             => _db.Deposits.Where(d => d.UserId == userId).ToList();
 
-        public bool ApproveTransaction(int txId)
+        public (ServiceResult Result, string Message) ApproveTransaction(int txId)
         {
             var transaction = _db.Deposits.FirstOrDefault(d => d.Id == txId);
             if(transaction is null)
-                return false;
+                return (ServiceResult.NotFound, "Transaction not found");
+            if (transaction.Status != DepositStatus.Undecided)
+                return (ServiceResult.NotValid, "Deposit status already decided");
 
             transaction.Status = DepositStatus.Approved;
             _userBalanceService.AddUserBalance(transaction.UserId, transaction.CurrencyId, transaction.Amount);
             _db.SaveChanges();
-            return true;
+            return (ServiceResult.Success, "");
         }
 
-        public bool DeclineTransaction(int txId)
+        public (ServiceResult Result, string Message) DeclineTransaction(int txId)
         {
-            var transaction = _db.Deposits.FirstOrDefault(d => d.Id == txId);
+            var transaction = GetTransactionById(txId);
             if (transaction is null)
-                return false;
+                return (ServiceResult.NotFound, "Transaction not found");
+            if (transaction.Status != DepositStatus.Undecided)
+                return (ServiceResult.NotValid, "Deposit status already decided");
+
 
             transaction.Status = DepositStatus.Declined;
             _db.SaveChanges();
-            return true;
+            return (ServiceResult.Success, "");
         }
+
+        public Deposit? GetTransactionById(int txId)
+            => _db.Deposits.FirstOrDefault(d => d.Id == txId);
     }
 }
