@@ -1,4 +1,4 @@
-﻿using UserWallet.Services.Extensions;
+﻿using System.Diagnostics.CodeAnalysis;
 
 namespace UserWallet.Controllers
 {
@@ -7,21 +7,23 @@ namespace UserWallet.Controllers
     public class AuthController : ControllerBase
     {
         private readonly IUserService _userService;
+        private readonly IAuthService _authService;
         private readonly IUserBalanceService _userBalanceService;
 
-        public AuthController(IUserService userService, IUserBalanceService userBalanceService)
+        public AuthController(IUserService userService, IUserBalanceService userBalanceService, IAuthService authService)
         {
             _userService = userService;
             _userBalanceService = userBalanceService;
+            _authService = authService;
         }
 
         [HttpPost("sign-up")]
-        public IActionResult Add([FromBody] SignInDTO userDto)
+        public IActionResult Add([FromBody] SignUpDTO userDto)
         {
             if(!ModelState.IsValid)
                 return BadRequest(ModelState.Values.SelectMany(v => v.Errors));
 
-            var result = _userService.AddUser(userDto.Username, userDto.Password);
+            var result = _userService.AddUser(userDto.Username!, userDto.Password!);
             if (!result)
                 return BadRequest("The user already exists");
 
@@ -31,24 +33,24 @@ namespace UserWallet.Controllers
         [HttpPost("login")]
         async public Task<IActionResult> Login([FromBody] LoginDTO userDto)
         {
-            User? user = _userService.GetUserByNameAndPassword(userDto.Username, userDto.Password);
+            if (HttpContext.GetCurrentUserId() is not null)
+                return BadRequest("The user is already logged in");
+
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState.Values.SelectMany(v => v.Errors));
+
+            User? user = _userService.GetUserByNameAndPassword(userDto.Username!, userDto.Password!);
             if (user is null)
                 return Unauthorized();
 
-            user.Balances = _userBalanceService.GetUserBalances(user.Id);
-
-            var claims = new List<Claim>
-            {
-                new Claim(ClaimTypes.Name, user.Username),
-                new Claim(ClaimTypes.Role, user.Role),
-                new Claim(ClaimTypes.NameIdentifier, user.Id.ToString())
-            };
-            ClaimsIdentity identity = new ClaimsIdentity(claims, "Cookies");
-            ClaimsPrincipal principal = new ClaimsPrincipal(identity);
+            user.Balances = _userBalanceService.GetUserBalances(user.Id).Balances;
+            var principal = _authService.MakeClaims(user);
             await HttpContext.SignInAsync(CookieAuthenticationDefaults.AuthenticationScheme, principal);
-            
+
             return Ok(user);
         }
+
+        
 
         [Authorize]
         [HttpPost("logout")]
@@ -60,12 +62,18 @@ namespace UserWallet.Controllers
 
         [HttpPatch("change-password")]
         [Authorize]
-        public IActionResult ChangePassword(string newPassword)
+        public IActionResult ChangePassword([FromBody] ChangeUserPasswordDTO model)
         {
             if(!ModelState.IsValid)
                 return BadRequest(ModelState.Values.SelectMany(v => v.Errors));
 
-            _userService.ChangePassword(HttpContext.GetCurrentUserId(), newPassword);
+            int userId = HttpContext.GetCurrentUserId()!.Value;
+
+            var user = _userService.GetUserById(userId);
+            if (user!.Password != model.OldPassword)
+                return BadRequest("Wrong old password");
+
+            _userService.ChangePassword(userId, model.NewPassword!);
             return Ok();
         }
     }
